Check record permissions on export
