+++ ukrdc_fastapi/routers/api/v1/patientrecords/pid/export.py
@@ -2,6 +2,7 @@ from fastapi import APIRouter, Depends, HTTPException, Security
 from mirth_client.exceptions import MirthPostError
 from mirth_client.mirth import MirthAPI
 from redis import Redis
+from ukrdc_sqla.ukrdc import PatientRecord
 
 from ukrdc_fastapi.dependencies import get_mirth, get_redis
 from ukrdc_fastapi.dependencies.audit import Auditer, RecordOperation, get_auditer
@@ -15,6 +16,8 @@ from ukrdc_fastapi.utils.mirth import (
     get_channel_from_name,
 )
 
+from .dependencies import _get_patientrecord
+
 router = APIRouter(tags=["Patient Records/Export"])
 
 
@@ -24,7 +27,7 @@ router = APIRouter(tags=["Patient Records/Export"])
     dependencies=[Security(auth.permission(Permissions.EXPORT_RECORDS))],
 )
 async def patient_export_pv(
-    pid: str,
+    patient_record: PatientRecord = Depends(_get_patientrecord),
     mirth: MirthAPI = Depends(get_mirth),
     redis: Redis = Depends(get_redis),
     audit: Auditer = Depends(get_auditer),
@@ -36,13 +39,13 @@ async def patient_export_pv(
             500, detail="ID for PV Outbound channel not found"
         )  # pragma: no cover
 
-    message: str = build_export_all_message(pid)
+    message: str = build_export_all_message(patient_record.pid)
     try:
         await channel.post_message(message)
     except MirthPostError as e:
         raise HTTPException(500, str(e)) from e  # pragma: no cover
 
-    audit.add_patient_record(pid, None, None, RecordOperation.EXPORT_PV)
+    audit.add_patient_record(patient_record.pid, None, None, RecordOperation.EXPORT_PV)
 
     return MirthMessageResponseSchema(status="success", message=message)
 
@@ -53,7 +56,7 @@ async def patient_export_pv(
     dependencies=[Security(auth.permission(Permissions.EXPORT_RECORDS))],
 )
 async def patient_export_pv_tests(
-    pid: str,
+    patient_record: PatientRecord = Depends(_get_patientrecord),
     mirth: MirthAPI = Depends(get_mirth),
     redis: Redis = Depends(get_redis),
     audit: Auditer = Depends(get_auditer),
@@ -65,13 +68,15 @@ async def patient_export_pv_tests(
             500, detail="ID for PV Outbound channel not found"
         )  # pragma: no cover
 
-    message: str = build_export_tests_message(pid)
+    message: str = build_export_tests_message(patient_record.pid)
     try:
         await channel.post_message(message)
     except MirthPostError as e:
         raise HTTPException(500, str(e)) from e  # pragma: no cover
 
-    audit.add_patient_record(pid, None, None, RecordOperation.EXPORT_PV_TESTS)
+    audit.add_patient_record(
+        patient_record.pid, None, None, RecordOperation.EXPORT_PV_TESTS
+    )
 
     return MirthMessageResponseSchema(status="success", message=message)
 
@@ -82,7 +87,7 @@ async def patient_export_pv_tests(
     dependencies=[Security(auth.permission(Permissions.EXPORT_RECORDS))],
 )
 async def patient_export_pv_docs(
-    pid: str,
+    patient_record: PatientRecord = Depends(_get_patientrecord),
     mirth: MirthAPI = Depends(get_mirth),
     redis: Redis = Depends(get_redis),
     audit: Auditer = Depends(get_auditer),
@@ -94,13 +99,15 @@ async def patient_export_pv_docs(
             500, detail="ID for PV Outbound channel not found"
         )  # pragma: no cover
 
-    message: str = build_export_docs_message(pid)
+    message: str = build_export_docs_message(patient_record.pid)
     try:
         await channel.post_message(message)
     except MirthPostError as e:
         raise HTTPException(500, str(e)) from e  # pragma: no cover
 
-    audit.add_patient_record(pid, None, None, RecordOperation.EXPORT_PV_DOCS)
+    audit.add_patient_record(
+        patient_record.pid, None, None, RecordOperation.EXPORT_PV_DOCS
+    )
 
     return MirthMessageResponseSchema(status="success", message=message)
 
@@ -111,7 +118,7 @@ async def patient_export_pv_docs(
     dependencies=[Security(auth.permission(Permissions.EXPORT_RECORDS))],
 )
 async def patient_export_radar(
-    pid: str,
+    patient_record: PatientRecord = Depends(_get_patientrecord),
     mirth: MirthAPI = Depends(get_mirth),
     redis: Redis = Depends(get_redis),
     audit: Auditer = Depends(get_auditer),
@@ -123,12 +130,14 @@ async def patient_export_radar(
             500, detail="ID for RADAR Outbound channel not found"
         )  # pragma: no cover
 
-    message: str = build_export_radar_message(pid)
+    message: str = build_export_radar_message(patient_record.pid)
     try:
         await channel.post_message(message)
     except MirthPostError as e:
         raise HTTPException(500, str(e)) from e  # pragma: no cover
 
-    audit.add_patient_record(pid, None, None, RecordOperation.EXPORT_RADAR)
+    audit.add_patient_record(
+        patient_record.pid, None, None, RecordOperation.EXPORT_RADAR
+    )
 
     return MirthMessageResponseSchema(status="success", message=message)
